pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crisis-commander-env"
version = "0.1.0"
description = "An OpenEnv-style urban crisis command environment for training high-stakes decision agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Punith Venkat Sai" }
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "openenv-core>=0.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.setuptools.packages.find]
include = ["crisis_commander_env*", "server*"]

